+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Fix to apply range penalty to AOE template placement. [#404](https://github.com/dmdorman/hero6e-foundryvtt/issues/404)
 
 # Version 3.0.42
-- Fix where previous actor migrations were incomplete.
+- Fix where previous actor migrations were incomplete. [#399](https://github.com/dmdorman/hero6e-foundryvtt/issues/399)
 - Full Health now resets charges.
 - Fix "Actor Description".
 - Improved Mental Defense description.